# backend/cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/native/aliyun/presign/download": {
            "post": {
                "description": "presign download aliyun oss object",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aliyun"
                ],
                "summary": "presign download aliyun oss object",
                "parameters": [
                    {
                        "description": "presign download request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg_web_services_native_aliyun.PresignDownloadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/example_com_nano_template_pkg_web_middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/pkg_web_services_native_aliyun.PresignResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/native/aliyun/presign/list": {
            "post": {
                "description": "presign list aliyun oss objects",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aliyun"
                ],
                "summary": "presign list aliyun oss objects",
                "parameters": [
                    {
                        "type": "string",
                        "description": "bucket prefix",
                        "name": "bucket_prefix",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "marker",
                        "name": "marker",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "description": "presign list request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg_web_services_native_aliyun.PresignListRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/example_com_nano_template_pkg_web_middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/pkg_web_services_native_aliyun.ListObjectsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/native/aliyun/presign/upload": {
            "post": {
                "description": "presign upload aliyun oss object",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "aliyun"
                ],
                "summary": "presign upload aliyun oss object",
                "parameters": [
                    {
                        "description": "presign upload request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pkg_web_services_native_aliyun.PresignUploadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/example_com_nano_template_pkg_web_middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/pkg_web_services_native_aliyun.PresignResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/template/{id}": {
            "get": {
                "description": "get template by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "common"
                ],
                "summary": "get template by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Template ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/example_com_nano_template_pkg_web_middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/pkg_web_services_common.GetTemplateResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "example_com_nano_template_pkg_config.TemplateFrontMatter": {
            "type": "object",
            "properties": {
                "role": {
                    "type": "string"
                }
            }
        },
...
        "example_com_nano_template_pkg_web_middleware.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "0 for success, 1 for failure, -1 for error",
                    "type": "integer"
                },
                "data": {
                    "description": "response data"
                },
                "message": {
                    "description": "response message",
                    "type": "string"
                }
            }
        },
        "pkg_web_services_common.GetTemplateResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "frontmatter": {
                    "$ref": "#/definitions/example_com_nano_template_pkg_config.TemplateFrontMatter"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "pkg_web_services_native_aliyun.CommonPrefix": {
            "type": "object",
            "properties": {
                "prefix": {
                    "type": "string"
                }
            }
        },
        "pkg_web_services_native_aliyun.ListObjectsResponse": {
            "type": "object",
            "properties": {
                "common_prefixes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pkg_web_services_native_aliyun.CommonPrefix"
                    }
                },
                "contents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pkg_web_services_native_aliyun.ObjectProperties"
                    }
                },
                "delimiter": {
                    "type": "string"
                },
                "encoding_type": {
                    "type": "string"
                },
                "is_truncated": {
                    "type": "boolean"
                },
                "marker": {
                    "type": "string"
                },
                "max_keys": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "next_marker": {
                    "type": "string"
                },
                "prefix": {
                    "type": "string"
                }
            }
        },
        "pkg_web_services_native_aliyun.ObjectOwner": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "pkg_web_services_native_aliyun.ObjectProperties": {
            "type": "object",
            "properties": {
                "etag": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "last_modified": {
                    "type": "string"
                },
                "owner": {
                    "$ref": "#/definitions/pkg_web_services_native_aliyun.ObjectOwner"
                },
                "restore_info": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "storage_class": {
                    "type": "string"
                },
                "transition_time": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "pkg_web_services_native_aliyun.PresignDownloadRequest": {
            "type": "object",
            "properties": {
                "bucket_prefix": {
                    "type": "string"
                },
                "getter": {
                    "description": "获取方, 作为标识字段, 无特殊含义",
                    "type": "string"
                },
                "object_key": {
                    "type": "string"
                }
            }
        },
        "pkg_web_services_native_aliyun.PresignListRequest": {
            "type": "object",
            "properties": {
                "getter": {
                    "description": "获取方, 作为标识字段, 无特殊含义",
                    "type": "string"
                }
            }
        },
        "pkg_web_services_native_aliyun.PresignResponse": {
            "type": "object",
            "properties": {
                "expiration": {
                    "type": "string"
                },
                "method": {
                    "type": "string"
                },
                "signed_headers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pkg_web_services_native_aliyun.SignedHeader"
                    }
                },
                "signed_url": {
                    "type": "string"
                }
            }
        },
        "pkg_web_services_native_aliyun.PresignUploadRequest": {
            "type": "object",
            "required": [
                "object_key"
            ],
            "properties": {
                "mime": {
                    "type": "string"
                },
                "object_key": {
                    "type": "string"
                },
                "sender": {
                    "description": "发送方, 作为标识字段, 无特殊含义",
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "pkg_web_services_native_aliyun.SignedHeader": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
    }
}